+++ CHANGELOG.md
@@ -15,8 +15,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Adds oihana\core\accessors\setKeyValue
 - Adds oihana\core\date\formatDateTime
 - Adds oihana\core\reflections\getFunctionInfo
+- Adds oihana\core\strings\block
+- Adds oihana\core\strings\blockPrefix
+- Adds oihana\core\strings\blockSuffix
 - Adds oihana\core\strings\formatFromDocument
 - Adds oihana\core\strings\isRegexp
+- Adds oihana\core\strings\wrapBlock
 
 ### Removed
 - Keep only the core folder !